Conflicts:
	.gitignore
	Addons/WCell.Collision/WorldMapTile.cs
	Addons/WCell.DefaultAddon/Spells/DeathKnight/DeathKnightBloodFixes.cs
	Addons/WCell.DefaultAddon/Spells/DeathKnight/DeathKnightDiseases.cs
	Addons/WCell.DefaultAddon/Spells/DeathKnight/DeathKnightFixes.cs
	Addons/WCell.DefaultAddon/Spells/DeathKnight/DeathKnightFrostFixes.cs
	Addons/WCell.DefaultAddon/Spells/DeathKnight/DeathKnightUnholyFixes.cs
	Addons/WCell.DefaultAddon/Spells/Druid/DruidBalanceFixes.cs
	Addons/WCell.DefaultAddon/Spells/Druid/DruidFeralCombatFixes.cs
	Addons/WCell.DefaultAddon/Spells/Druid/DruidRestorationFixes.cs
	Addons/WCell.DefaultAddon/Spells/Hunter/HunterFixes.cs
	Addons/WCell.DefaultAddon/Spells/Mage/MageFireFixes.cs
	Addons/WCell.DefaultAddon/Spells/MiscSpells.cs
	Addons/WCell.DefaultAddon/Spells/Paladin/PaladinFixes.cs
	Addons/WCell.DefaultAddon/Spells/Paladin/PaladinHolyFixes.cs
	Addons/WCell.DefaultAddon/Spells/Paladin/PaladinProtectionFixes.cs
	Addons/WCell.DefaultAddon/Spells/Paladin/PaladinRetributionFixes.cs
	Addons/WCell.DefaultAddon/Spells/Priest/PriestHolyFixes.cs
	Addons/WCell.DefaultAddon/Spells/Priest/PriestShadowFixes.cs
	Addons/WCell.DefaultAddon/Spells/Rogue/RogueFixes.cs
	Addons/WCell.DefaultAddon/Spells/Warrior/WarriorFuryFixes.cs
	Addons/WCell.DefaultAddon/Spells/Warrior/WarriorProtectionFixes.cs
	Applications/WCell.RealmServerConsole/Program.cs
	Applications/WCell.RealmServerConsole/RealmServerConsole.cs
	Core/WCell.Constants/Constants.Units.cs
	Core/WCell.Constants/Misc.cs
	Core/WCell.Constants/Movement.cs
	Core/WCell.Constants/OpCodes.cs
	Core/WCell.Constants/Spells/SpellId.cs
	Core/WCell.Constants/Spells/SpellLineId.cs
	Core/WCell.Constants/Spells/SpellMechanic.cs
	Core/WCell.Core/ClientInformation.cs
	Run/RealmServerAddons/DefaultAddon/DefaultAddonConfig.xml
	Services/WCell.RealmServer/Battlegrounds/BattlemasterList.cs
	Services/WCell.RealmServer/Battlegrounds/PvPDifficulty.cs
	Services/WCell.RealmServer/Battlegrounds/WorldSafeLocation.cs
	Services/WCell.RealmServer/Entities/Unit.Combat.cs
	Services/WCell.RealmServer/Entities/Unit.Mechanics.cs
	Services/WCell.RealmServer/Factions/FactionDBC.cs
	Services/WCell.RealmServer/Global/World.DBC.cs
	Services/WCell.RealmServer/Global/WorldEventMgr.cs
	Services/WCell.RealmServer/Global/Zone.cs
	Services/WCell.RealmServer/Handlers/MovementHandler.cs
	Services/WCell.RealmServer/Handlers/QueryHandler.cs
	Services/WCell.RealmServer/Handlers/QuestHandler.cs
	Services/WCell.RealmServer/Handlers/TalentHandler.cs
	Services/WCell.RealmServer/Items/Enchanting/EnchantMgr.cs
	Services/WCell.RealmServer/NPCs/NPCEntry.cs
	Services/WCell.RealmServer/NPCs/NPCMgr.cs
	Services/WCell.RealmServer/Quests/QuestHonor.cs
	Services/WCell.RealmServer/Quests/QuestRewRep.cs
	Services/WCell.RealmServer/Quests/QuestTemplate.cs
	Services/WCell.RealmServer/Quests/QuestXp.cs
	Services/WCell.RealmServer/RacesClasses/Archetypes.DBC.cs
	Services/WCell.RealmServer/Skills/SkillHandler.cs
	Services/WCell.RealmServer/Spells/Effects/AddProficiencyHandler.cs
	Services/WCell.RealmServer/Spells/PlayerSpellCollection.cs
	Services/WCell.RealmServer/Spells/Spell.Aura.cs
	Services/WCell.RealmServer/Spells/Spell.DBC.cs
	Services/WCell.RealmServer/Spells/Spell.Fields.cs
	Services/WCell.RealmServer/Spells/Spell.cs
	Services/WCell.RealmServer/Spells/SpellCast.Perform.cs
	Services/WCell.RealmServer/Spells/SpellCast.cs
	Services/WCell.RealmServer/Spells/SpellCollection.cs
	Services/WCell.RealmServer/Spells/SpellEffect.Fields.cs
	Services/WCell.RealmServer/Spells/SpellEffect.cs
	Services/WCell.RealmServer/Spells/SpellHandler.cs
	Services/WCell.RealmServer/Spells/SpellLines.Def.cs
	Services/WCell.RealmServer/Spells/SpellProcEventEntry.cs
	Services/WCell.RealmServer/Spells/SpellSummons.cs
	Services/WCell.RealmServer/Talents/Glyphs.DBC.cs
	Services/WCell.RealmServer/Talents/TalentMgr.cs
	Services/WCell.RealmServer/WCell.RealmServer.csproj
	Utilities/WCell.PacketAnalyzer/Logs/LogConverter.cs
	Utilities/WCell.PacketAnalyzer/WCell.PacketAnalyzer.csproj
	Utilities/WCell.Tools/Commands/RalekSpellStudiesCommand.cs
	Utilities/WCell.Tools/Domi/Instances.cs
	Utilities/WCell.Tools/Domi/WCellEnumWriter.cs
	Utilities/WCell.Tools/Domi/WorldMapOverlay.DBC.cs
	Utilities/WCell.Tools/Maps/Structures/DBCMapEntry.cs
	Utilities/WCell.Tools/Ralek/SpellStudies.cs
	Utilities/WCell.Tools/Spells/SpellFocusObject.DBC.cs
	Utilities/WCell.Tools/Spells/SpellLineWriter.cs
	WCell.sln

+++ Services/WCell.RealmServer/Skills/Skill.cs
@@ -4,7 +4,7 @@
  *   copyright		: (C) The WCell Team
  *   email		: info@wcell.org
  *   last changed	: $LastChangedDate: 2010-01-28 05:18:24 +0100 (to, 28 jan 2010) $
- *   last author	: $LastChangedBy: dominikseifert $
+
  *   revision		: $Rev: 1229 $
  *
  *   This program is free software; you can redistribute it and/or modify
@@ -15,7 +15,6 @@
  *************************************************************************/
 
 using System;
-using NLog;
 using WCell.Constants.Achievements;
 using WCell.Constants.Skills;
 using WCell.Constants.Updates;
@@ -206,7 +205,7 @@ namespace WCell.RealmServer.Skills
 			if (SkillLine.HasTier(tier))
 			{
 				uint tierLimit = SkillLine.Tiers.GetMaxValue(tier);
-				if (CurrentValue >= tierLimit - 100)
+				if (CurrentValue >= (int)tierLimit - 100)
 				{
 					// cannot be learnt if we have less than max - 100 in that skill
 					return true;
